Move cluster-agent AD bootstrap and profiling setup out of cmd/agent/common
Resolves the FIXME in cluster-agent's start command about importing
cmd/agent/common just for LoadComponents/SetupInternalProfiling. Moves
LoadComponents to pkg/config/autodiscovery and SetupInternalProfiling
to pkg/config/settings, so the cluster-agent binary no longer depends
on the main agent binary's internal package.
