+++ cmd/agent/common/loader.go
@@ -6,23 +6,13 @@
 package common
 
 import (
-	"path/filepath"
-
 	autodiscovery "github.com/DataDog/datadog-agent/comp/core/autodiscovery/def"
 	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/pkg/util/defaultpaths"
+	confad "github.com/DataDog/datadog-agent/pkg/config/autodiscovery"
 )
 
 // LoadComponents configures several common Agent components:
 // tagger, collector, scheduler and autodiscovery
 func LoadComponents(ac autodiscovery.Component, config config.Component) {
-	confdPath := config.GetString("confd_path")
-
-	confSearchPaths := []string{
-		confdPath,
-		filepath.Join(defaultpaths.GetDistPath(), "conf.d"),
-		"",
-	}
-
-	setupAutoDiscovery(confSearchPaths, ac, config)
+	confad.LoadComponents(ac, config)
 }
